Support for git-checkout of a specific config
Added support for checking out a specific branch, tag, or commit of the
configuration space if the configuration source is a Git repository.
The Git ref is specified in the FAI_CONFIG_SRC_REF variable, which can
be passed to FAI in the command line.

After checking out the FAI_CONFIG_SRC_REF, we re-run
$FAI/hooks/subroutines since the configuration may have changed.