Introduced the NOT_VZ_CONTAINER class
The NOT_VZ_CONTAINER class applies when FAI is not running in an OpenVZ
container, and therefore has to deal with kernels, kernel modules, and
bootloaders.

Renamed hooks/dirinstall.DEFAULT.source to
hooks/dirinstall.NOT_VZ_CONTAINER so that we can get rid of a
gatekeeping condition inside the script.