Friendly identifier of the configuration git ref
When checking out a particular ref of the configuration from Git, print
a human-friendly identifier of the commit being used.  If it is on the
master branch, produce a Subversion-like revision number by counting the
number of commits in the history ("master r123").  If it is not
obviously on the master branch, use `git describe --all` instead.