Fixed ARP for vertx[12].iun.ubc.ca
On mlookup1 and xlookup1, moved dns[12].iun.ubc.ca from the vlan1510 to
the lo interface as host-scoped addresses, and changed the arp_ignore
sysctl setting for the vlan1510 interface from 8 ("ignore all ARP on
that interface") to 3 ("ignore ARP for host-scoped addresses").  This
allows vertx[12].iun.ubc.ca, which remain on the vlan1510 interfaces, to
ARP normally.