Use khost1 disk config for all VM_HOSTs
This disk partitioning scheme is meant for a smaller first physical
drive (or logical hardware RAID volume) on which the host OS will
reside, and a huge second physical drive (or logical hardward RAID
volume) which will contain the virtual disk images of the VM guests.

Added the preserve_lazy option for the hostdata-vz logical volume.