Disable AcceptEnv in NFS root's sshd_config
Some programs (notably Perl) will complain when invoked with environment
variables corresponding to an unsupported locale.  Since the NFS root is
a minimalistic installation that is omits many locale definitions, it
makes sense to disable the "AcceptEnv LANG LC_*" line in the NFS root's
sshd_config to prevent these warnings.