Name network interfaces with ascending MAC addr
Rewrite udev rules such that network interfaces with consecutive (or
nearly consecutive) MAC addresses are named in ascending order.
Built-in Ethernet devices on motherboards usually have consecutive MAC
addresses, and the order of their MAC addresses generally coincides with
the intended order of enumeration.  However, udev will write its rules
file enumerating them according to their position in the topology of the
PCI bus.

For at least one Supermicro motherboard and one Dell machine, udev
enumerates the two Ethernet interfaces in reverse order; this commit
fixes it.

Some machines, such as IBM blades, allow the enumeration order of the
Ethernet interfaces to be configurable in the BIOS; this commit probably
breaks that feature.