Install freeipmi-tools, ipmitool if IPMI exists
If it appears that FAI is installing to a machine with IPMI support,
install the freeipmi-tools and ipmitool packages.

The freeipmi-tools package includes the bmc-info command, which can be
used by cfengine to probe for further information.  Note that bmc-info
can stall for 20 seconds if IPMI is not present, so we don't want to
install/run it indiscriminately.

If the machine is used in a Proxmox VE cluster, the fence_ipmilan agent
(included in the fence-agents-pve package) can use ipmitool to fence the
machine if it loses contact with the rest of the cluster.

The freeipmi-tools and ipmitool packages, with their dependencies, add
about 8 MB to the installation size.