Nicer preview of setup-storage actions
setup-storage, with no parameters, already does a pretty good job of
summarizing the actions to be taken.

One problem is if running setup-storage on a virgin disk.  When it tries
to run `parted -s` to discover what is currently on the disk, it will
say:

(STDOUT) Error: /dev/XXX: unrecognised disk label

But that's not really an error, so we downgrade the output to an "INFO"
message.