Allow fai-pxe -r COMMIT_ID
fai-pxe was being too strict when checking whether the git reference
specified by the -r option actually exists on the remote.  As all
easy verification techniques can only detect whether a branch or
tag name exists on the remote, and not whether a given commit ID
exists, I have downgraded the apparent non-existence from an error
to a warning.