Generalized disk configuration classes
The UBUNTU_32G class was simultaneously too rigid (the fai-setup-storage
configuration not good for anything besides a single disk of 32 GB) and too
broadly applied (it was chosen for all Ubuntu installations, regardless of the
number and sizes of disks present).

This change renames the UBUNTU_32G class to VIRTS, to be applied to any virtual
server with one disk.  To generalize the concept, virtual servers with two
disks will use the second disk for persistent storage -- a configuration
previously called DATABASE.

The class/80-disk script has been reorganized a bit.  Obsolete disk
configurations for msql1 and xsql1 have been removed.