Prevent last apt-get -f install from hanging
scripts/LAST/50-misc runs `apt-get -f install` as a final check to ensure that
packages are correctly installed and configured.  However, apt-get could
ask for confirmation on some operations, waiting forever for input from its
standard input.  An example would be if it needs to uninstall a package to
satisfy an inter-package constraint.

The solution is to close its standard input so that it cannot expect input
there.  Furthermore, using the --no-remove flag makes it more explicit that
we prefer not to approve unusual operations.