Removed Google Earth package
The make-googleearth-package command produces a .deb that has depends on
i386 libraries, even on amd64 installations.  That means that we would need
to first run `dpkg --add-architecture i386` to enable i386 as a foreign
architecture on an amd64 OS.  Furthermore, after installing the generated
googleearth package, we would need to run `apt-get -f install` to pull in
nearly 100 MB of i386 libraries.

I'm not convinced that there is sufficient demand for Google Earth to justify
this complication, so I am removing it until someone complains.