Install libnss-ldapd and libpam-ldapd
Install libnss-ldapd and libpam-ldapd instead of relying on CFengine to do it
post-installation.  Otherwise, for UBUNTU_STANDARD, the ldap-auth-client will
pull in the libnss-ldap and libpam-ldap packages as dependencies instead of
their modern replacements libnss-ldapd and libpam-ldapd.

On Ubuntu 14.04, in particular, libnss-ldap 264-2.2ubuntu4 triggers
Ubuntu bug 1350393, which causes accounts-daemon to hang indefinitely
unless dbus is restarted, due to a race condition.  As a result,

    /usr/bin/dbus-send --system --print-reply \
        --dest=org.freedesktop.Accounts \
        /org/freedesktop/Accounts org.freedesktop.Accounts.ListCachedUsers

will fail, and lightdm will take a long time to scan every home directory
for face icons before displaying the login screen.