Removed install-squeeze and rescue-squeeze
One NFS root can cross-install another distribution or version of Linux.
Therefore, the Linux distro to be installed should be specified as a
class rather than hard-coded in the PXE template.

Removed the install-squeeze PXE template, in favour of the new command
  SERVER-SIDE/bin/fai-pxe -t install -c DEBIAN_6 machine.ece.ubc.ca

Removed the rescue-squeeze PXE template, in favour of the new command
  SERVER-SIDE/bin/fai-pxe -t rescue machine.ece.ubc.ca