Partition empty disks
If setup-storage encounters a disk to be partitioned, but the disk
contains no partition table at all, it will give up.  To prevent such
failures, this hook creates an empty partition table of the type
specified in the setup-storage configuration file on any disk whose
first 512 bytes are all null.