Make DHCPC class active unless NO_DHCPC is given
DHCPC is the more common case, so it should be the default.
Note that if there is no /etc/network/interfaces file, modern Linux
distros will use NetworkManager, which is mostly fine, but it's
slower to activate at boot, which can cause race conditions with
LDAP and/or automount.

This change also places bci04, mehdi-pc02, kramer, and jerry in the
DHCPC class (previously excluded by accidental omission).