CENTOS GRUB fix
Use a more sophisticated way to determine which disk(s) to install GRUB
on.  Don't assume that if /dev/md0 exists then we want to install GRUB
to the first two disks, and that if it's not a DISK2 installation then
we want to install to the first disk.  Rather, try to determine the
physical disk(s) on which the boot partition resides (unless the DISK2
class applies, in which case we install to the first two disks).

Always install lvm2 on CENTOS_7, since all of our partitioning schemes
use it.  Without it, dracut would generate an initrd without LVM
support.