Rewrote dirinstall hook for fstab manipulation
FAI 3.4 has a bug in /usr/bin/fai-do-scripts that causes it to skip Perl
scripts.  (Case sensitivity problem: it accepts "perl" scripts, but the
`file` command reports that it is a "Perl" script.)

So, I've rewritten it as a shell script that uses awk.  Also, don't
hard-code "/altroot" in the script.