Use dialog for partitioning preview
Show a confirmation dialog on the system console with a preview of the
partitioning actions.  Remote access is also possible by sending a
signal to the dialog process: SIGUSR1 to proceed, SIGUSR2 to open a
shell on the console, or SIGTERM to abort.  Confirmation is implied if
no input is received in two hours.