Use ncurses dialog when pausing before reboot
dialog --pause can give a countdown and an easy way to terminate the
pause from the console, or launch a shell.

Also changed the partitioning preview dialog to require the FAI createvt
flag to show the "Shell" button.  The FAI menu flag alone should not be
sufficient, due to security concerns.