EFI System Partition and GRUB partition
Use FAI's gpt-bios partitioning support.  On all GPT-partitioned disks
processed by setup-storage, we will have both an EFI system partition
and a BIOS boot partition.  This will allow the same partition table to
be used both for EFI and for BIOS booting.

Previously, scripts/GRUB_PC/50-setup would run parted to convert
partition 1 of every disk into a BIOS boot partition.  This is bad,
because it makes an assumption about the disk configuration that could
be inconsistent with the disk_config that is actually being used by
setup-storage.

FAI's standard way to create a BIOS boot partition is to specify
"gpt-bios" as the partition table type in the disk_config file.  This
will cause setup-storage to append a BIOS boot partition at the start of
the disk (out of order).  However, there may not be enough space to
create this BIOS boot partition if the machine already has, say,
partition 1 being an EFI system partition at the start of the disk, and
the disk_config specifies that partition 1 is to be preserved.

If the disk already has a BIOS boot partition as partition 1 (from FAI
installations under the old scheme), we can simply remove it, and let it
be re-created by gpt-bios.