Try to detect an EFI system partition
GRUB_EFI requires an EFI system partition.  However, some FAI disk
configs might not specify an EFI system partition, especially if Linux
is being installed to a second disk.  So, we find the first partition
that has the right partition type UUID and that contains a FAT
filesystem, and make an fstab entry for it.