Removed VM_GUEST partitioning scheme
The VM_GUEST partitioning scheme was based on the standard virtual
machine images provided by UBC IT Systems.  It led to some headaches,
though:

- Lack of consistency within ECE.  If we test some setup on a virtual
  machine, it should behave similarly when we deploy it to a physical
  machine.

- Lack of a /data partition for the single-disk case.  It's useful to
  have space to store data that persists across OS reinstallation, or
  that is at least easily identifiable as valuable data.

- Lack of an EFI system partition.  (With an EFI system partition added,
  an advantage of unification went away.)

- Useless /var/log/audit partition, in Ubuntu.

The VM_GUEST2 is renamed to DATA2, to emphasize the use of the second
disk for /data, and to de-emphasize its association with virtualization.
Its first disk now has /altroot support.