Removed most of sample FAISERVER configuration
We don't need to run an APT proxy, since UBC IT has a local mirror.

We don't want to create a private 192.168.1.0/24 network.  We have a
separate DHCP server.

/etc/default/tftpd-hpa is managed by cfengine.

We don't need to run fai-setup.  The NFS root is built using the
configuration and makefile in SERVER-SIDE/nfsroots/ of this repository.
Exporting the NFS root is done by cfengine.  FAI logs are transferred
over HTTPS rather than over SSH/SCP/RSH.  The FAI configuration is
served from a Git repository.

We don't want to remove /etc/resolv.conf, even if the FAI server is a
DHCP client; that breaks when /etc/resolv.conf is managed by resolvconf.