Also install stretch-backports kernel in NFS root
In the NFS root, install the stretch-backports kernel alongside the
stock stretch kernel.  The backports kernel may be needed to support new
hardware, but is not as well tested as the stock kernel.  Invoking
fai-pxe will configure a machine to netboot the stock kernel; edit the
PXELINUX configuration file manually if the backports kernel is desired.

The stretch-backports kernel is installed via a hook: listing
linux-image-amd64/stretch-backports would cause linux-image-amd64 to be
upgraded from stock to backports, rather than installing a second
kernel.

The hooks had to be reordered, such that the release-id script (which
defines $ARCH) is run before the backport-kernel script (which installs
the second kernel).