Ignore removable disks when counting disks
If using a USB stick to assist with PXE booting, for example, then it
could inadvertently trigger a "Too many hard disks.  Too risky." error
in class/80-disk.

This might slightly increase the risk of accidentally trying to install
Linux onto the USB stick, if the USB stick is assigned a device name
that comes before the intended installation target disk, but that is
unlikely in most bus topologies.  In any case, the USB stick would
likely not have enough space, so partitioning would not succeed.