Added debconf response for package wireshark-common to debconf/OMNI
Dumpcap can be installed in a way that allows members of the "wireshark"
system group to capture packets. This is recommended over the
alternative of running Wireshark/Tshark directly as root, because
less of the code will run with elevated privileges.

As ECE does not restrict access to network interfaces, debconf response is set to "true", allowing members of the "wireshark" group to capture packets.