Obtain FQDN using reverse-DNS lookup
Previously, we assumed that the fully qualified domain name of the target
machine should be "$HOSTNAME.$DOMAIN".  This is incorrect.  The $DOMAIN,
which derives from the "domain name" DHCP option provided by the DHCP server,
is supposed to specify the implicit suffix when performing DNS lookups; it
may have nothing to do with the FQDN of the target host.

Unfortunately, the dhclient-script(8) in the initrd, which interprets the
DHCP offer, does not receive the client FQDN (DHCP option 81, RFC 4702).
Therefore, the next best solution is to perform a reverse-DNS lookup on the IP
address, and use it if the unqualified hostnames match.