Don't remove /etc/hostname
It was previously thought that DHCP clients don't need /etc/hostname,
because the DHCP server would provide the hostname, and not hard-coding
the hostname in the filesystem would be handy when generating an image
for repeated or cloud deployment.

However, removing /etc/hostname turns out to be a bad idea because DHCP
does not provide the FQDN, and the FQDN cannot always be correctly
inferred by joining the short hostname and the DNS search domain.

Newer versions of the Linux hostname package (>= 3.13) say this in the
hostname(1) man page:

/etc/hostname: Historically this file was supposed to only contain the
       hostname and not the full canonical FQDN. Nowadays most software
       is able to cope with a full FQDN here. This file is read at boot
       time by the system initialization scripts to set the hostname.