Use LVM for swap; Enlarge ESP
Instead of using a partition for swap space, use a logical volume.
Rationale:

1. Years ago, it was advisable to use a contiguous space for swap, and
to place it near the beginning of the disk, due to performance concerns.
Contiguity is no longer an issue with SSDs, which are more common now.
Furthermore, swapping is rarer with the modern availability of RAM.

2. LVs, unlike partitions, are resizable.  Modern Linux distributions
can use swap space for hibernation, if it is larger than physical
memory, which makes resizability more desirable.

3. Modern Linux distributions' installers would generally use LVM for
swap.

To allow the PV that contains the /data filesystem to be preserved, we
need to maintain the old partition numbering, so that FAI's
setup-storage does not get confused.  This means that we need to create
a dummy partition table entry where the old swap partition entry used to
be.

Unfortunately, when using FAI to reinstall, the former swap space would
be wasted.  Furthermore, setup-storage may fail to find enough space
within the preserved VG to make a swap LV, in which case manual
intervention is needed.

As long as we are in the business of creating dummy partitions, we might
as well get rid of the DISK1P3 scheme, and use DISK1P5 instead, since
the only difference is inserting a few more dummy partitions.

---

In addition, we will now prefer to create a larger EFI system partition,
since the Boot Loader Specification
(https://systemd.io/BOOT_LOADER_SPECIFICATION/) says:

    "If the OS is installed on a disk with GPT disk label, and an
    ESP partition (i.e. with the GPT type UID of
    c12a7328-f81f-11d2-ba4b-00a0c93ec93b) already exists and is
    large enough (let’s say 250MB) and otherwise qualifies, it
    should be used as $BOOT."

We might as well be more generous with the EFI system partition size,
for future-proofing.