Updated Ubuntu 20.04 package list
Removed packages that are no longer in the distribution (notably all
Python 2 packages), as well as packages that are redundant with what is
in debootstrap.

Tasks are now metapackages.

Removed Chromium browser: in Ubuntu 20.04, the chromium-browser dpkg
tries to install a snap.  This currently leads to a timeout error during
installation.

Avoid letting lubuntu-desktop uninstall grub-efi: lubuntu-desktop
20.04.6 depends on lubuntu-grub-theme 20.04.3, which recommends grub2,
which conflicts with grub-efi.  Hopefully that will be fixed soon, but
in the meantime, a workaround is to avoid installing packages that are
transitively recommended by lubuntu-desktop.

Removed resolvconf from Ubuntu 20.04: With systemd-resolved and
network-manager, resolvconf seems to produce empty
/etc/resolv.conf files.

ntp superseded by systemd-timesyncd